Incorporate a full source-map library for PhantomJS into the headless runner.
Split up the slimer/phantom runners to that this code only applies to
Phantom.

This makes the phantom.js a total mess and isn't really a suitable
solution for adding source maps, but it demos what's possible. Ideally
the source map code would be pulled out as its own Phantom-requirable
library but I haven't had time to set that up. The biggest problem is
that the mapstrace library requires the source-map library, which itself
doesn't import easily into Phantom.
