+++ library/resources/runners/phantom.js
@@ -1,0 +1,1352 @@
+// Headless Script testing
+// Phantomjs script for running cljs.test tests
+//
+var fs = require('fs');
+
+/** To print source-mapped stacktraces for Phantom, some code from
+ * the `mapstrace` and `source-map` npm libraries, with suitable
+ * modifictions to flatten exports inline, is inlined below.
+ *
+ * Ideally this would be brought in via libraries but it's a giant 
+ * pain with Phantom's require behavior and I haven't managed to make
+ * it work.
+ *
+ * mapstrace: https://github.com/janekp/mapstrace (MIT License)
+ *
+ * source-map: https://github.com/mozilla/source-map (New BSD License)
+ *
+ */
+
+/*** source-map/base64.js ****/
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ */
+
+var base64 = {};
+var charToIntMap = {};
+var intToCharMap = {};
+
+'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/'
+.split('')
+.forEach(function (ch, index) {
+    charToIntMap[ch] = index;
+    intToCharMap[index] = ch;
+});
+
+/**
+ * Encode an integer in the range of 0 to 63 to a single base 64 digit.
+ */
+base64.encode = function base64_encode(aNumber) {
+    if (aNumber in intToCharMap) {
+        return intToCharMap[aNumber];
+    }
+    throw new TypeError("Must be between 0 and 63: " + aNumber);
+};
+
+/**
+ * Decode a single base 64 digit to an integer.
+ */
+base64.decode = function base64_decode(aChar) {
+    if (aChar in charToIntMap) {
+        return charToIntMap[aChar];
+    }
+    throw new TypeError("Not a valid base 64 digit: " + aChar);
+};
+
+
+/*** source-map/base64-vlq.js ****/
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Based on the Base 64 VLQ implementation in Closure Compiler:
+ * https://code.google.com/p/closure-compiler/source/browse/trunk/src/com/google/debugging/sourcemap/Base64VLQ.java
+ *
+ * Copyright 2011 The Closure Compiler Authors. All rights reserved.
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are
+ * met:
+ *
+ *  * Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *  * Redistributions in binary form must reproduce the above
+ *    copyright notice, this list of conditions and the following
+ *    disclaimer in the documentation and/or other materials provided
+ *    with the distribution.
+ *  * Neither the name of Google Inc. nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+var base64VLQ = {};
+
+// A single base 64 digit can contain 6 bits of data. For the base 64 variable
+// length quantities we use in the source map spec, the first bit is the sign,
+// the next four bits are the actual value, and the 6th bit is the
+// continuation bit. The continuation bit tells us whether there are more
+// digits in this value following this digit.
+//
+//   Continuation
+//   |    Sign
+//   |    |
+//   V    V
+//   101011
+
+var VLQ_BASE_SHIFT = 5;
+
+// binary: 100000
+var VLQ_BASE = 1 << VLQ_BASE_SHIFT;
+
+// binary: 011111
+var VLQ_BASE_MASK = VLQ_BASE - 1;
+
+// binary: 100000
+var VLQ_CONTINUATION_BIT = VLQ_BASE;
+
+/**
+ * Converts from a two-complement value to a value where the sign bit is
+ * is placed in the least significant bit.  For example, as decimals:
+ *   1 becomes 2 (10 binary), -1 becomes 3 (11 binary)
+ *   2 becomes 4 (100 binary), -2 becomes 5 (101 binary)
+ */
+function toVLQSigned(aValue) {
+    return aValue < 0
+        ? ((-aValue) << 1) + 1
+        : (aValue << 1) + 0;
+}
+
+/**
+ * Converts to a two-complement value from a value where the sign bit is
+ * is placed in the least significant bit.  For example, as decimals:
+ *   2 (10 binary) becomes 1, 3 (11 binary) becomes -1
+ *   4 (100 binary) becomes 2, 5 (101 binary) becomes -2
+ */
+function fromVLQSigned(aValue) {
+    var isNegative = (aValue & 1) === 1;
+    var shifted = aValue >> 1;
+    return isNegative
+        ? -shifted
+        : shifted;
+}
+
+/**
+ * Returns the base 64 VLQ encoded value.
+ */
+base64VLQ.encode = function base64VLQ_encode(aValue) {
+    var encoded = "";
+    var digit;
+
+    var vlq = toVLQSigned(aValue);
+
+    do {
+        digit = vlq & VLQ_BASE_MASK;
+        vlq >>>= VLQ_BASE_SHIFT;
+        if (vlq > 0) {
+            // There are still more digits in this value, so we must make sure the
+            // continuation bit is marked.
+            digit |= VLQ_CONTINUATION_BIT;
+        }
+        encoded += base64.encode(digit);
+    } while (vlq > 0);
+
+    return encoded;
+};
+
+/**
+ * Decodes the next base 64 VLQ value from the given string and returns the
+ * value and the rest of the string.
+ */
+base64VLQ.decode = function base64VLQ_decode(aStr) {
+    var i = 0;
+    var strLen = aStr.length;
+    var result = 0;
+    var shift = 0;
+    var continuation, digit;
+
+    do {
+        if (i >= strLen) {
+            throw new Error("Expected more digits in base 64 VLQ value.");
+        }
+        digit = base64.decode(aStr.charAt(i++));
+        continuation = !!(digit & VLQ_CONTINUATION_BIT);
+        digit &= VLQ_BASE_MASK;
+        result = result + (digit << shift);
+        shift += VLQ_BASE_SHIFT;
+    } while (continuation);
+
+    return {
+        value: fromVLQSigned(result),
+            rest: aStr.slice(i)
+    };
+};
+
+
+/*** source-map/binary-search.js ***/
+
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ */
+
+var binarySearch = {};
+
+/**
+ * Recursive implementation of binary search.
+ *
+ * @param aLow Indices here and lower do not contain the needle.
+ * @param aHigh Indices here and higher do not contain the needle.
+ * @param aNeedle The element being searched for.
+ * @param aHaystack The non-empty array being searched.
+ * @param aCompare Function which takes two elements and returns -1, 0, or 1.
+ */
+function recursiveSearch(aLow, aHigh, aNeedle, aHaystack, aCompare) {
+    // This function terminates when one of the following is true:
+    //
+    //   1. We find the exact element we are looking for.
+    //
+    //   2. We did not find the exact element, but we can return the next
+    //      closest element that is less than that element.
+    //
+    //   3. We did not find the exact element, and there is no next-closest
+    //      element which is less than the one we are searching for, so we
+    //      return null.
+    var mid = Math.floor((aHigh - aLow) / 2) + aLow;
+    var cmp = aCompare(aNeedle, aHaystack[mid], true);
+    if (cmp === 0) {
+        // Found the element we are looking for.
+        return aHaystack[mid];
+    }
+    else if (cmp > 0) {
+        // aHaystack[mid] is greater than our needle.
+        if (aHigh - mid > 1) {
+            // The element is in the upper half.
+            return recursiveSearch(mid, aHigh, aNeedle, aHaystack, aCompare);
+        }
+        // We did not find an exact match, return the next closest one
+        // (termination case 2).
+        return aHaystack[mid];
+    }
+    else {
+        // aHaystack[mid] is less than our needle.
+        if (mid - aLow > 1) {
+            // The element is in the lower half.
+            return recursiveSearch(aLow, mid, aNeedle, aHaystack, aCompare);
+        }
+        // The exact needle element was not found in this haystack. Determine if
+        // we are in termination case (2) or (3) and return the appropriate thing.
+        return aLow < 0
+            ? null
+            : aHaystack[aLow];
+    }
+}
+
+/**
+ * This is an implementation of binary search which will always try and return
+ * the next lowest value checked if there is no exact hit. This is because
+ * mappings between original and generated line/col pairs are single points,
+ * and there is an implicit region between each of them, so a miss just means
+ * that you aren't on the very start of a region.
+ *
+ * @param aNeedle The element you are looking for.
+ * @param aHaystack The array that is being searched.
+ * @param aCompare A function which takes the needle and an element in the
+ *     array and returns -1, 0, or 1 depending on whether the needle is less
+ *     than, equal to, or greater than the element, respectively.
+ */
+binarySearch.search = function search(aNeedle, aHaystack, aCompare) {
+    return aHaystack.length > 0
+        ? recursiveSearch(-1, aHaystack.length, aNeedle, aHaystack, aCompare)
+        : null;
+};
+
+
+/*** source-map/util.js ***/
+
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ */
+var util = {};
+
+/**
+ * This is a helper function for getting values from parameter/options
+ * objects.
+ *
+ * @param args The object we are extracting values from
+ * @param name The name of the property we are getting.
+ * @param defaultValue An optional value to return if the property is missing
+ * from the object. If this is not specified and the property is missing, an
+ * error will be thrown.
+ */
+function getArg(aArgs, aName, aDefaultValue) {
+    if (aName in aArgs) {
+        return aArgs[aName];
+    } else if (arguments.length === 3) {
+        return aDefaultValue;
+    } else {
+        throw new Error('"' + aName + '" is a required argument.');
+    }
+}
+util.getArg = getArg;
+
+var urlRegexp = /([\w+\-.]+):\/\/((\w+:\w+)@)?([\w.]+)?(:(\d+))?(\S+)?/;
+var dataUrlRegexp = /^data:.+\,.+/;
+
+function urlParse(aUrl) {
+    var match = aUrl.match(urlRegexp);
+    if (!match) {
+        return null;
+    }
+    return {
+        scheme: match[1],
+            auth: match[3],
+            host: match[4],
+            port: match[6],
+            path: match[7]
+    };
+}
+util.urlParse = urlParse;
+
+function urlGenerate(aParsedUrl) {
+    var url = aParsedUrl.scheme + "://";
+    if (aParsedUrl.auth) {
+        url += aParsedUrl.auth + "@"
+    }
+    if (aParsedUrl.host) {
+        url += aParsedUrl.host;
+    }
+    if (aParsedUrl.port) {
+        url += ":" + aParsedUrl.port
+    }
+    if (aParsedUrl.path) {
+        url += aParsedUrl.path;
+    }
+    return url;
+}
+util.urlGenerate = urlGenerate;
+
+function join(aRoot, aPath) {
+    var url;
+
+    if (aPath.match(urlRegexp) || aPath.match(dataUrlRegexp)) {
+        return aPath;
+    }
+
+    if (aPath.charAt(0) === '/' && (url = urlParse(aRoot))) {
+        url.path = aPath;
+        return urlGenerate(url);
+    }
+
+    return aRoot.replace(/\/$/, '') + '/' + aPath;
+}
+util.join = join;
+
+/**
+ * Because behavior goes wacky when you set `__proto__` on objects, we
+ * have to prefix all the strings in our set with an arbitrary character.
+ *
+ * See https://github.com/mozilla/source-map/pull/31 and
+ * https://github.com/mozilla/source-map/issues/30
+ *
+ * @param String aStr
+ */
+function toSetString(aStr) {
+    return '$' + aStr;
+}
+util.toSetString = toSetString;
+
+function fromSetString(aStr) {
+    return aStr.substr(1);
+}
+util.fromSetString = fromSetString;
+
+function relative(aRoot, aPath) {
+    aRoot = aRoot.replace(/\/$/, '');
+
+    var url = urlParse(aRoot);
+    if (aPath.charAt(0) == "/" && url && url.path == "/") {
+        return aPath.slice(1);
+    }
+
+    return aPath.indexOf(aRoot + '/') === 0
+        ? aPath.substr(aRoot.length + 1)
+        : aPath;
+}
+util.relative = relative;
+
+function strcmp(aStr1, aStr2) {
+    var s1 = aStr1 || "";
+    var s2 = aStr2 || "";
+    return (s1 > s2) - (s1 < s2);
+}
+
+/**
+ * Comparator between two mappings where the original positions are compared.
+ *
+ * Optionally pass in `true` as `onlyCompareGenerated` to consider two
+ * mappings with the same original source/line/column, but different generated
+ * line and column the same. Useful when searching for a mapping with a
+ * stubbed out mapping.
+ */
+function compareByOriginalPositions(mappingA, mappingB, onlyCompareOriginal) {
+    var cmp;
+
+    cmp = strcmp(mappingA.source, mappingB.source);
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.originalLine - mappingB.originalLine;
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.originalColumn - mappingB.originalColumn;
+    if (cmp || onlyCompareOriginal) {
+        return cmp;
+    }
+
+    cmp = strcmp(mappingA.name, mappingB.name);
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.generatedLine - mappingB.generatedLine;
+    if (cmp) {
+        return cmp;
+    }
+
+    return mappingA.generatedColumn - mappingB.generatedColumn;
+};
+util.compareByOriginalPositions = compareByOriginalPositions;
+
+/**
+ * Comparator between two mappings where the generated positions are
+ * compared.
+ *
+ * Optionally pass in `true` as `onlyCompareGenerated` to consider two
+ * mappings with the same generated line and column, but different
+ * source/name/original line and column the same. Useful when searching for a
+ * mapping with a stubbed out mapping.
+ */
+function compareByGeneratedPositions(mappingA, mappingB, onlyCompareGenerated) {
+    var cmp;
+
+    cmp = mappingA.generatedLine - mappingB.generatedLine;
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.generatedColumn - mappingB.generatedColumn;
+    if (cmp || onlyCompareGenerated) {
+        return cmp;
+    }
+
+    cmp = strcmp(mappingA.source, mappingB.source);
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.originalLine - mappingB.originalLine;
+    if (cmp) {
+        return cmp;
+    }
+
+    cmp = mappingA.originalColumn - mappingB.originalColumn;
+    if (cmp) {
+        return cmp;
+    }
+
+    return strcmp(mappingA.name, mappingB.name);
+};
+util.compareByGeneratedPositions = compareByGeneratedPositions;
+
+
+/*** source-map/array-set.js ***/
+
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ */
+
+/**
+ * A data structure which is a combination of an array and a set. Adding a new
+ * member is O(1), testing for membership is O(1), and finding the index of an
+ * element is O(1). Removing elements from the set is not supported. Only
+ * strings are supported for membership.
+ */
+function ArraySet() {
+    this._array = [];
+    this._set = {};
+}
+
+/**
+ * Static method for creating ArraySet instances from an existing array.
+ */
+ArraySet.fromArray = function ArraySet_fromArray(aArray, aAllowDuplicates) {
+    var set = new ArraySet();
+    for (var i = 0, len = aArray.length; i < len; i++) {
+        set.add(aArray[i], aAllowDuplicates);
+    }
+    return set;
+};
+
+/**
+ * Add the given string to this set.
+ *
+ * @param String aStr
+ */
+ArraySet.prototype.add = function ArraySet_add(aStr, aAllowDuplicates) {
+    var isDuplicate = this.has(aStr);
+    var idx = this._array.length;
+    if (!isDuplicate || aAllowDuplicates) {
+        this._array.push(aStr);
+    }
+    if (!isDuplicate) {
+        this._set[util.toSetString(aStr)] = idx;
+    }
+};
+
+/**
+ * Is the given string a member of this set?
+ *
+ * @param String aStr
+ */
+ArraySet.prototype.has = function ArraySet_has(aStr) {
+    return Object.prototype.hasOwnProperty.call(this._set,
+            util.toSetString(aStr));
+};
+
+/**
+ * What is the index of the given string in the array?
+ *
+ * @param String aStr
+ */
+ArraySet.prototype.indexOf = function ArraySet_indexOf(aStr) {
+    if (this.has(aStr)) {
+        return this._set[util.toSetString(aStr)];
+    }
+    throw new Error('"' + aStr + '" is not in the set.');
+};
+
+/**
+ * What is the element at the given index?
+ *
+ * @param Number aIdx
+ */
+ArraySet.prototype.at = function ArraySet_at(aIdx) {
+    if (aIdx >= 0 && aIdx < this._array.length) {
+        return this._array[aIdx];
+    }
+    throw new Error('No element indexed by ' + aIdx);
+};
+
+/**
+ * Returns the array representation of this set (which has the proper indices
+ * indicated by indexOf). Note that this is a copy of the internal array used
+ * for storing the members so that no one can mess with internal state.
+ */
+ArraySet.prototype.toArray = function ArraySet_toArray() {
+    return this._array.slice();
+};
+
+
+/*** source-map/source-map-consumer.js ***/
+
+/*
+ * Copyright 2011 Mozilla Foundation and contributors
+ * Licensed under the New BSD license. See LICENSE or:
+ * http://opensource.org/licenses/BSD-3-Clause
+ */
+
+/**
+ * A SourceMapConsumer instance represents a parsed source map which we can
+ * query for information about the original file positions by giving it a file
+ * position in the generated source.
+ *
+ * The only parameter is the raw source map (either as a JSON string, or
+ * already parsed to an object). According to the spec, source maps have the
+ * following attributes:
+ *
+ *   - version: Which version of the source map spec this map is following.
+ *   - sources: An array of URLs to the original source files.
+ *   - names: An array of identifiers which can be referrenced by individual mappings.
+ *   - sourceRoot: Optional. The URL root from which all sources are relative.
+ *   - sourcesContent: Optional. An array of contents of the original source files.
+ *   - mappings: A string of base64 VLQs which contain the actual mappings.
+ *   - file: The generated file this source map is associated with.
+ *
+ * Here is an example source map, taken from the source map spec[0]:
+ *
+ *     {
+ *       version : 3,
+ *       file: "out.js",
+ *       sourceRoot : "",
+ *       sources: ["foo.js", "bar.js"],
+ *       names: ["src", "maps", "are", "fun"],
+ *       mappings: "AA,AB;;ABCDE;"
+ *     }
+ *
+ * [0]: https://docs.google.com/document/d/1U1RGAehQwRypUTovF1KRlpiOFze0b-_2gc6fAH0KY0k/edit?pli=1#
+ */
+function SourceMapConsumer(aSourceMap) {
+    var sourceMap = aSourceMap;
+    if (typeof aSourceMap === 'string') {
+        sourceMap = JSON.parse(aSourceMap.replace(/^\)\]\}'/, ''));
+    }
+
+    var version = util.getArg(sourceMap, 'version');
+    var sources = util.getArg(sourceMap, 'sources');
+    // Sass 3.3 leaves out the 'names' array, so we deviate from the spec (which
+    // requires the array) to play nice here.
+    var names = util.getArg(sourceMap, 'names', []);
+    var sourceRoot = util.getArg(sourceMap, 'sourceRoot', null);
+    var sourcesContent = util.getArg(sourceMap, 'sourcesContent', null);
+    var mappings = util.getArg(sourceMap, 'mappings');
+    var file = util.getArg(sourceMap, 'file', null);
+
+    // Once again, Sass deviates from the spec and supplies the version as a
+    // string rather than a number, so we use loose equality checking here.
+    if (version != this._version) {
+        throw new Error('Unsupported version: ' + version);
+    }
+
+    // Pass `true` below to allow duplicate names and sources. While source maps
+    // are intended to be compressed and deduplicated, the TypeScript compiler
+    // sometimes generates source maps with duplicates in them. See Github issue
+    // #72 and bugzil.la/889492.
+    this._names = ArraySet.fromArray(names, true);
+    this._sources = ArraySet.fromArray(sources, true);
+
+    this.sourceRoot = sourceRoot;
+    this.sourcesContent = sourcesContent;
+    this._mappings = mappings;
+    this.file = file;
+}
+
+/**
+ * Create a SourceMapConsumer from a SourceMapGenerator.
+ *
+ * @param SourceMapGenerator aSourceMap
+ *        The source map that will be consumed.
+ * @returns SourceMapConsumer
+ */
+SourceMapConsumer.fromSourceMap =
+function SourceMapConsumer_fromSourceMap(aSourceMap) {
+    var smc = Object.create(SourceMapConsumer.prototype);
+
+    smc._names = ArraySet.fromArray(aSourceMap._names.toArray(), true);
+    smc._sources = ArraySet.fromArray(aSourceMap._sources.toArray(), true);
+    smc.sourceRoot = aSourceMap._sourceRoot;
+    smc.sourcesContent = aSourceMap._generateSourcesContent(smc._sources.toArray(),
+            smc.sourceRoot);
+    smc.file = aSourceMap._file;
+
+    smc.__generatedMappings = aSourceMap._mappings.slice()
+        .sort(util.compareByGeneratedPositions);
+    smc.__originalMappings = aSourceMap._mappings.slice()
+        .sort(util.compareByOriginalPositions);
+
+    return smc;
+};
+
+/**
+ * The version of the source mapping spec that we are consuming.
+ */
+SourceMapConsumer.prototype._version = 3;
+
+/**
+ * The list of original sources.
+ */
+Object.defineProperty(SourceMapConsumer.prototype, 'sources', {
+    get: function () {
+             return this._sources.toArray().map(function (s) {
+                 return this.sourceRoot ? util.join(this.sourceRoot, s) : s;
+             }, this);
+         }
+});
+
+// `__generatedMappings` and `__originalMappings` are arrays that hold the
+// parsed mapping coordinates from the source map's "mappings" attribute. They
+// are lazily instantiated, accessed via the `_generatedMappings` and
+// `_originalMappings` getters respectively, and we only parse the mappings
+// and create these arrays once queried for a source location. We jump through
+// these hoops because there can be many thousands of mappings, and parsing
+// them is expensive, so we only want to do it if we must.
+//
+// Each object in the arrays is of the form:
+//
+//     {
+//       generatedLine: The line number in the generated code,
+//       generatedColumn: The column number in the generated code,
+//       source: The path to the original source file that generated this
+//               chunk of code,
+//       originalLine: The line number in the original source that
+//                     corresponds to this chunk of generated code,
+//       originalColumn: The column number in the original source that
+//                       corresponds to this chunk of generated code,
+//       name: The name of the original symbol which generated this chunk of
+//             code.
+//     }
+//
+// All properties except for `generatedLine` and `generatedColumn` can be
+// `null`.
+//
+// `_generatedMappings` is ordered by the generated positions.
+//
+// `_originalMappings` is ordered by the original positions.
+
+SourceMapConsumer.prototype.__generatedMappings = null;
+Object.defineProperty(SourceMapConsumer.prototype, '_generatedMappings', {
+    get: function () {
+             if (!this.__generatedMappings) {
+                 this.__generatedMappings = [];
+                 this.__originalMappings = [];
+                 this._parseMappings(this._mappings, this.sourceRoot);
+             }
+
+             return this.__generatedMappings;
+         }
+});
+
+SourceMapConsumer.prototype.__originalMappings = null;
+Object.defineProperty(SourceMapConsumer.prototype, '_originalMappings', {
+    get: function () {
+             if (!this.__originalMappings) {
+                 this.__generatedMappings = [];
+                 this.__originalMappings = [];
+                 this._parseMappings(this._mappings, this.sourceRoot);
+             }
+
+             return this.__originalMappings;
+         }
+});
+
+/**
+ * Parse the mappings in a string in to a data structure which we can easily
+ * query (the ordered arrays in the `this.__generatedMappings` and
+ * `this.__originalMappings` properties).
+ */
+SourceMapConsumer.prototype._parseMappings =
+function SourceMapConsumer_parseMappings(aStr, aSourceRoot) {
+    var generatedLine = 1;
+    var previousGeneratedColumn = 0;
+    var previousOriginalLine = 0;
+    var previousOriginalColumn = 0;
+    var previousSource = 0;
+    var previousName = 0;
+    var mappingSeparator = /^[,;]/;
+    var str = aStr;
+    var mapping;
+    var temp;
+
+    while (str.length > 0) {
+        if (str.charAt(0) === ';') {
+            generatedLine++;
+            str = str.slice(1);
+            previousGeneratedColumn = 0;
+        }
+        else if (str.charAt(0) === ',') {
+            str = str.slice(1);
+        }
+        else {
+            mapping = {};
+            mapping.generatedLine = generatedLine;
+
+            // Generated column.
+            temp = base64VLQ.decode(str);
+            mapping.generatedColumn = previousGeneratedColumn + temp.value;
+            previousGeneratedColumn = mapping.generatedColumn;
+            str = temp.rest;
+
+            if (str.length > 0 && !mappingSeparator.test(str.charAt(0))) {
+                // Original source.
+                temp = base64VLQ.decode(str);
+                mapping.source = this._sources.at(previousSource + temp.value);
+                previousSource += temp.value;
+                str = temp.rest;
+                if (str.length === 0 || mappingSeparator.test(str.charAt(0))) {
+                    throw new Error('Found a source, but no line and column');
+                }
+
+                // Original line.
+                temp = base64VLQ.decode(str);
+                mapping.originalLine = previousOriginalLine + temp.value;
+                previousOriginalLine = mapping.originalLine;
+                // Lines are stored 0-based
+                mapping.originalLine += 1;
+                str = temp.rest;
+                if (str.length === 0 || mappingSeparator.test(str.charAt(0))) {
+                    throw new Error('Found a source and line, but no column');
+                }
+
+                // Original column.
+                temp = base64VLQ.decode(str);
+                mapping.originalColumn = previousOriginalColumn + temp.value;
+                previousOriginalColumn = mapping.originalColumn;
+                str = temp.rest;
+
+                if (str.length > 0 && !mappingSeparator.test(str.charAt(0))) {
+                    // Original name.
+                    temp = base64VLQ.decode(str);
+                    mapping.name = this._names.at(previousName + temp.value);
+                    previousName += temp.value;
+                    str = temp.rest;
+                }
+            }
+
+            this.__generatedMappings.push(mapping);
+            if (typeof mapping.originalLine === 'number') {
+                this.__originalMappings.push(mapping);
+            }
+        }
+    }
+
+    this.__originalMappings.sort(util.compareByOriginalPositions);
+};
+
+/**
+ * Find the mapping that best matches the hypothetical "needle" mapping that
+ * we are searching for in the given "haystack" of mappings.
+ */
+SourceMapConsumer.prototype._findMapping =
+function SourceMapConsumer_findMapping(aNeedle, aMappings, aLineName,
+        aColumnName, aComparator) {
+    // To return the position we are searching for, we must first find the
+    // mapping for the given position and then return the opposite position it
+    // points to. Because the mappings are sorted, we can use binary search to
+    // find the best mapping.
+
+    if (aNeedle[aLineName] <= 0) {
+        throw new TypeError('Line must be greater than or equal to 1, got '
+                + aNeedle[aLineName]);
+    }
+    if (aNeedle[aColumnName] < 0) {
+        throw new TypeError('Column must be greater than or equal to 0, got '
+                + aNeedle[aColumnName]);
+    }
+
+    return binarySearch.search(aNeedle, aMappings, aComparator);
+};
+
+/**
+ * Returns the original source, line, and column information for the generated
+ * source's line and column positions provided. The only argument is an object
+ * with the following properties:
+ *
+ *   - line: The line number in the generated source.
+ *   - column: The column number in the generated source.
+ *
+ * and an object is returned with the following properties:
+ *
+ *   - source: The original source file, or null.
+ *   - line: The line number in the original source, or null.
+ *   - column: The column number in the original source, or null.
+ *   - name: The original identifier, or null.
+ */
+SourceMapConsumer.prototype.originalPositionFor =
+function SourceMapConsumer_originalPositionFor(aArgs) {
+    var needle = {
+        generatedLine: util.getArg(aArgs, 'line'),
+        generatedColumn: util.getArg(aArgs, 'column')
+    };
+
+    var mapping = this._findMapping(needle,
+            this._generatedMappings,
+            "generatedLine",
+            "generatedColumn",
+            util.compareByGeneratedPositions);
+
+    if (mapping) {
+        var source = util.getArg(mapping, 'source', null);
+        if (source && this.sourceRoot) {
+            source = util.join(this.sourceRoot, source);
+        }
+        return {
+            source: source,
+                line: util.getArg(mapping, 'originalLine', null),
+                column: util.getArg(mapping, 'originalColumn', null),
+                name: util.getArg(mapping, 'name', null)
+        };
+    }
+
+    return {
+        source: null,
+            line: null,
+            column: null,
+            name: null
+    };
+};
+
+/**
+ * Returns the original source content. The only argument is the url of the
+ * original source file. Returns null if no original source content is
+ * availible.
+ */
+SourceMapConsumer.prototype.sourceContentFor =
+function SourceMapConsumer_sourceContentFor(aSource) {
+    if (!this.sourcesContent) {
+        return null;
+    }
+
+    if (this.sourceRoot) {
+        aSource = util.relative(this.sourceRoot, aSource);
+    }
+
+    if (this._sources.has(aSource)) {
+        return this.sourcesContent[this._sources.indexOf(aSource)];
+    }
+
+    var url;
+    if (this.sourceRoot
+            && (url = util.urlParse(this.sourceRoot))) {
+        // XXX: file:// URIs and absolute paths lead to unexpected behavior for
+        // many users. We can help them out when they expect file:// URIs to
+        // behave like it would if they were running a local HTTP server. See
+        // https://bugzilla.mozilla.org/show_bug.cgi?id=885597.
+        var fileUriAbsPath = aSource.replace(/^file:\/\//, "");
+        if (url.scheme == "file"
+                && this._sources.has(fileUriAbsPath)) {
+            return this.sourcesContent[this._sources.indexOf(fileUriAbsPath)]
+        }
+
+        if ((!url.path || url.path == "/")
+                && this._sources.has("/" + aSource)) {
+            return this.sourcesContent[this._sources.indexOf("/" + aSource)];
+        }
+    }
+
+    throw new Error('"' + aSource + '" is not in the SourceMap.');
+};
+
+/**
+ * Returns the generated line and column information for the original source,
+ * line, and column positions provided. The only argument is an object with
+ * the following properties:
+ *
+ *   - source: The filename of the original source.
+ *   - line: The line number in the original source.
+ *   - column: The column number in the original source.
+ *
+ * and an object is returned with the following properties:
+ *
+ *   - line: The line number in the generated source, or null.
+ *   - column: The column number in the generated source, or null.
+ */
+SourceMapConsumer.prototype.generatedPositionFor =
+function SourceMapConsumer_generatedPositionFor(aArgs) {
+    var needle = {
+        source: util.getArg(aArgs, 'source'),
+        originalLine: util.getArg(aArgs, 'line'),
+        originalColumn: util.getArg(aArgs, 'column')
+    };
+
+    if (this.sourceRoot) {
+        needle.source = util.relative(this.sourceRoot, needle.source);
+    }
+
+    var mapping = this._findMapping(needle,
+            this._originalMappings,
+            "originalLine",
+            "originalColumn",
+            util.compareByOriginalPositions);
+
+    if (mapping) {
+        return {
+            line: util.getArg(mapping, 'generatedLine', null),
+                column: util.getArg(mapping, 'generatedColumn', null)
+        };
+    }
+
+    return {
+        line: null,
+            column: null
+    };
+};
+
+SourceMapConsumer.GENERATED_ORDER = 1;
+SourceMapConsumer.ORIGINAL_ORDER = 2;
+
+/**
+ * Iterate over each mapping between an original source/line/column and a
+ * generated line/column in this source map.
+ *
+ * @param Function aCallback
+ *        The function that is called with each mapping.
+ * @param Object aContext
+ *        Optional. If specified, this object will be the value of `this` every
+ *        time that `aCallback` is called.
+ * @param aOrder
+ *        Either `SourceMapConsumer.GENERATED_ORDER` or
+ *        `SourceMapConsumer.ORIGINAL_ORDER`. Specifies whether you want to
+ *        iterate over the mappings sorted by the generated file's line/column
+ *        order or the original's source/line/column order, respectively. Defaults to
+ *        `SourceMapConsumer.GENERATED_ORDER`.
+ */
+SourceMapConsumer.prototype.eachMapping =
+function SourceMapConsumer_eachMapping(aCallback, aContext, aOrder) {
+    var context = aContext || null;
+    var order = aOrder || SourceMapConsumer.GENERATED_ORDER;
+
+    var mappings;
+    switch (order) {
+        case SourceMapConsumer.GENERATED_ORDER:
+            mappings = this._generatedMappings;
+            break;
+        case SourceMapConsumer.ORIGINAL_ORDER:
+            mappings = this._originalMappings;
+            break;
+        default:
+            throw new Error("Unknown order of iteration.");
+    }
+
+    var sourceRoot = this.sourceRoot;
+    mappings.map(function (mapping) {
+        var source = mapping.source;
+        if (source && sourceRoot) {
+            source = util.join(sourceRoot, source);
+        }
+        return {
+            source: source,
+        generatedLine: mapping.generatedLine,
+        generatedColumn: mapping.generatedColumn,
+        originalLine: mapping.originalLine,
+        originalColumn: mapping.originalColumn,
+        name: mapping.name
+        };
+    }).forEach(aCallback, context);
+};
+
+var sourceMap = {"SourceMapConsumer": SourceMapConsumer};
+
+
+/*** based on mapstrace/mapstrace.js ***/
+
+/**
+ * Significantly modified to work with the stack trace output from
+ * PhantomJS's WebKit onError handler.
+ *
+ */
+
+var path = {"sep": fs.separator};
+
+function build(trace, full, fn) {
+    var result = [ ];
+    var files = { };
+    var i, c, t, f;
+    var src = { };
+
+    function _build() {
+        var unk = false;
+        var r = 0;
+        
+        for(i = 0, c = trace.length; i < c; i++) {
+            t = trace[i];
+            var fileName = t.file || t.sourceURL;
+            fileName = fileName.replace("file://", "");
+            
+            if(files[fileName].map) {
+                // QtWebKit throws JS errors without column numbers! This screws up the source map!
+                // If the output is one-statement per line, this seems to work, though.
+                t = files[fileName].map.originalPositionFor({ line: t.line, column: 1000000});
+                result.push(t);
+
+                unk = false;
+                
+                if(full && t.source && !t.name && !src[t.source]) {
+                    src[t.source] = { };
+                    r += 1;
+                }
+            } else if(!unk) {
+                result.push(1);
+                unk = true;
+            } else {
+                result[result.length - 1] = result[result.length - 1] + 1;
+            }
+        }
+        
+        if(r > 0) {
+            i = 0;
+
+            console.log("Looking up srcs " + JSON.stringify(src));
+            
+            for(f in src) {
+                if(src.hasOwnProperty(f)) {
+                    _readSrc(f);
+                }
+            }
+        } else {
+            fn(result);
+        }
+    }
+    
+    function _buildNames() {
+        var s;
+        
+        for(i = 0, c = result.length; i < c; i++) {
+            t = result[i];
+            
+            if(typeof(t) !== 'number') {
+                if(!t.name && t.source) {
+                    s = src[t.source];
+                    
+                    if(s && s.lines && t.line >= 1 && t.line <= s.lines.length) {
+                        t.name = "'" + s.lines[t.line - 1].replace(/^\s\s*/, '').replace(/[;\s]\s*$/, '') + "'";
+                        
+                        if(t.name.indexOf("'//") === 0) {
+                            t.line += 1;
+                            t.name = null;
+                            i -= 1;
+                        }
+                    }
+                }
+            }
+        }
+        
+        fn(result);
+    }
+    
+    function _readSrc(f) {
+        i++;
+
+        console.log("grabbing file " + f);
+        
+        var data;
+        try {
+            data = fs.read(f);
+        } catch (err) {
+            console.log("Couldn't read source file " + f);
+        }
+
+        if(data) {
+            src[f].lines = data.split("\n");
+        }
+
+        i--;
+
+        if(i === 0) {
+            _buildNames();
+        }
+
+        console.log("done with " + f);
+    }
+    
+    function _readMap(f) {
+        i++;
+
+        if (fs.isFile(f + '.map')) {
+            var data = fs.read(f + ".map");
+            if(data) {
+                try {
+                    files[f].map = new sourceMap.SourceMapConsumer(data.toString());
+                }
+                catch(e) {
+                    console.log('Could not parse "' + f + '.map" -- ' + e);
+                }
+            }
+
+            i--;
+
+            if(i === 0) {
+                _build();
+            }
+            if(i === 0) {
+                _build();
+            }
+        } else {
+            i--;
+        }
+    }
+    
+    if(typeof(full) === 'function') {
+        fn = full;
+        full = false;
+    }
+    
+    for(i = 0, c = trace.length; i < c; i++) {
+        t = trace[i];
+        var fileName = t.file;
+        if (!fileName) {
+            fileName = t.sourceURL
+        }
+        fileName = fileName.replace("file://", "")
+        
+        if(!files[fileName]) {
+            files[fileName] = { };
+        }
+    }
+    
+    i = 0;
+    
+    for(f in files) {
+        if(files.hasOwnProperty(f) &&
+            f !== 'module.js' && f !== 'node.js' && f !== 'http.js' && f !== 'events.js') {
+            _readMap(f);
+        }
+    }
+    
+    if(i === 0) {
+        fn(null);
+    }
+}
+
+function stringify(data, prefix, limit) {
+    var result = '';
+    var i, c, t;
+    
+    if(prefix && prefix.length > 0 && prefix.charAt(prefix.length - 1) != path.sep) {
+        prefix += path.sep;
+    }
+    
+    if(!limit) {
+        // /a/b/c/d/e/f/g -> .../c/d/e/f/g
+        limit = 5;
+    }
+    
+    if(data && data.length > 0) {
+        for(i = 0, c = data.length; i < c; i++) {
+            t = data[i];
+            
+            if(typeof(t) === 'number') {
+                result += (t > 1) ? '    ... (omitted ' + t + ' rows)' : '    ...';
+            } else if(t.source === null) {
+                result += '???';
+            } else {
+                if(prefix && t.source.indexOf(prefix) === 0) {
+                    t.source = t.source.substr(prefix.length);
+                } else {
+                    t.parts = t.source.split(path.sep);
+                    
+                    if(t.parts.length > limit) {
+                        t.source = '...' + path.sep + t.parts.slice(t.parts.length - limit).join(path.sep);
+                    }
+                }
+                
+                result += "    at " + ((t.name !== null) ? t.name : '???') + ' (' + t.source + ':' + t.line + ':' + t.column + ')';
+            }
+            
+            result += "\n";
+        }
+    }
+    
+    return result;
+}
+
+var mapstrace = {"build": build, "stringify": stringify};
+
+
+/*** back to headless.js ***/
+
+
+var p = require('webpage').create();
+var sys = require('system');
+var os = sys.os;
+
+// phantomjs has problems with windows drive letters, it thinks them to be a protocol definition. By using
+// triple slashes with file: protocol definition it works as expected
+function fixFileProtocol(path) {
+    return os.name === 'windows' ? "file:///" + path : "file://" + path;
+}
+
+var pagePath = fs.absolute("doo-index.html");
+
+var scripts = "";
+
+for (var i = 1; i < sys.args.length; i++) {
+    scripts += '<script src="' + fixFileProtocol(fs.absolute(sys.args[i])) + '"></script>';
+}
+
+var html = "<html><head><meta charset=\"UTF-8\">"
+           + scripts
+           + "</head><body></body></html>";
+
+fs.write(pagePath, html, 'w');
+
+function exit(code) {
+    phantom.exit(code);
+}
+
+// Prepare the page with triggers
+
+// Redirects the output of the page into the script
+p.onConsoleMessage = function(msg) {
+    console.log(msg);
+};
+
+p.onError = function(msg, trace) {
+    console.error(msg);
+    mapstrace.build(trace, false, function(result) {
+        console.log(mapstrace.stringify(result));
+    });
+    exit(1);
+};
+
+p.open(fixFileProtocol(pagePath), function (status) {
+    fs.remove(pagePath);
+    if (status == "fail") {
+        console.log(noScriptMsg("Phantom"));
+        exit(1);
+    } else {
+
+        p.onCallback = function (x) {
+	    var line = x.toString();
+	    if (line !== "[NEWLINE]") {
+	        console.log(line.replace(/\[NEWLINE\]/g, "\n"));
+	    }
+        };
+
+        // p.evaluate is sandboxed, can't ship closures across;
+        // so, a bit of a hack, better than polling :-P
+        var exitCodePrefix = "phantom-exit-code:";
+        p.onAlert = function (msg) {
+	    var exitCode = msg.replace(exitCodePrefix, "");
+	    if (msg != exitCode) {
+                exit(parseInt(exitCode));
+            }
+        };
+
+        p.evaluate(function(hasDoo, noDooMsg, exitCodePrefix) {
+            // Helper funciton to exit
+            function exit(isSuccess) {
+                window.alert(exitCodePrefix + (isSuccess ? 0 : 1));
+	    }
+            if (hasDoo()) {
+                // If doo is present, set up printing and exit, and start!
+	        doo.runner.set_print_fn_BANG_(function(x) {
+	            // using callPhantom to work around
+                    // https://github.com/laurentj/slimerjs/issues/223
+	            window.callPhantom(x.replace(/\n/g, "[NEWLINE]"));
+                    // since console.log *itself* adds a newline
+	        });
+	        doo.runner.set_exit_point_BANG_(exit);
+
+                // Start the tests, which should call the exit
+                // function when they are done.
+                var results = doo.runner.run_BANG_();
+
+            } else {
+                // If doo is not present, present an error and exit
+                window.callPhantom(noDooMsg);
+                exit(false);
+            }
+        }, hasDoo, noDooMsg(), exitCodePrefix);
+    }
+});
